Blink Detection
A system that detects faces and eyes with Haar_cascade and then uses a counter system to detect if the eyes are not detected for a certain period of time as a blink and prints it on the screen. The average person's blinking time is between 0.1 and 0.5 seconds. I took this time as 0.4 seconds in this code because I did not want any momentary eye detection error to be detected as an eye blink.